Clean up to the cross platform support in DirectX plugin.

+++ src/osgPlugins/directx/directx.h
@@ -24,6 +24,8 @@
 #ifndef _DIRECTX_H_
 #define _DIRECTX_H_
 
+#include <osg/Math>
+
 #include <string>
 #include <vector>
 #include <fstream>
@@ -40,7 +42,7 @@ namespace DX {
         float x,y,z;
 
         inline void normalize() {
-            float lenRecip = 1.0f / (float) sqrt(x * x + y * y + z * z);
+            float lenRecip = 1.0f / sqrtf(x * x + y * y + z * z);
             x *= lenRecip;
             y *= lenRecip;
             z *= lenRecip;
